refactor: extract out ViewRolesWindow and ViewPermissions component
